refactor/AnimatedView 모듈화
