+++ src/screens/Login/NicknameWrite/index.tsx
@@ -1,19 +1,13 @@
 import CameraIcon from '@assets/svgs/camera.svg';
+import AnimatedView from '@components/atom/AnimatedView';
 import BG from '@components/atom/BG';
 import Button from '@components/atom/Button';
 import DismissKeyboardView from '@components/atom/DismissKeyboardView';
 import Txt from '@components/atom/Txt';
 import {NativeStackScreenProps} from '@react-navigation/native-stack';
 import {AuthStackParamList} from '@stackNav/Auth';
-import {useEffect, useRef, useState} from 'react';
-import {
-  Animated,
-  Image,
-  Keyboard,
-  Pressable,
-  TextInput,
-  View,
-} from 'react-native';
+import {useEffect, useState} from 'react';
+import {Image, Keyboard, Pressable, TextInput, View} from 'react-native';
 import {
   ImageLibraryOptions,
   ImagePickerResponse,
@@ -47,9 +41,6 @@ const NicknameWriteScreen = ({route, navigation}: Readonly<AuthProps>) => {
   const [clickedUpload, setClickedUpload] = useState(false);
   const [isKeyboardVisible, setIsKeyboardVisible] = useState(false);
 
-  const fadeAnim = useRef(new Animated.Value(0)).current;
-  const translateYAnim = useRef(new Animated.Value(50)).current;
-
   useEffect(() => {
     const showSubscription = Keyboard.addListener('keyboardDidShow', () =>
       setIsKeyboardVisible(true),
@@ -65,36 +56,6 @@ const NicknameWriteScreen = ({route, navigation}: Readonly<AuthProps>) => {
       hideSubscription.remove();
     };
   }, []);
-
-  useEffect(() => {
-    if (clickedUpload) {
-      Animated.parallel([
-        Animated.timing(fadeAnim, {
-          toValue: 1,
-          duration: 200,
-          useNativeDriver: true,
-        }),
-        Animated.timing(translateYAnim, {
-          toValue: 0,
-          duration: 200,
-          useNativeDriver: true,
-        }),
-      ]).start();
-    } else {
-      Animated.parallel([
-        Animated.timing(fadeAnim, {
-          toValue: 0,
-          duration: 200,
-          useNativeDriver: true,
-        }),
-        Animated.timing(translateYAnim, {
-          toValue: 50,
-          duration: 200,
-          useNativeDriver: true,
-        }),
-      ]).start();
-    }
-  }, [clickedUpload]);
 
   const selectImage = async () => {
     const options: ImageLibraryOptions = {
@@ -217,12 +178,9 @@ const NicknameWriteScreen = ({route, navigation}: Readonly<AuthProps>) => {
           }`}>
           {/* 내부 컴포넌트에는 상위 onPress 이벤트가 전파되지 않도록 함 */}
           <Pressable onPress={() => {}} className="w-full">
-            <Animated.View
-              style={{
-                opacity: fadeAnim,
-                transform: [{translateY: translateYAnim}],
-                borderRadius: 10,
-              }}
+            <AnimatedView
+              visible={clickedUpload}
+              style={{borderRadius: 10}}
               className="bg-blue500 mb-[24]">
               <View className="h-[43] justify-center items-center">
                 <Txt
@@ -251,7 +209,7 @@ const NicknameWriteScreen = ({route, navigation}: Readonly<AuthProps>) => {
                   className="text-white"
                 />
               </Pressable>
-            </Animated.View>
+            </AnimatedView>
 
             <Button text="취소" onPress={() => setClickedUpload(false)} />
           </Pressable>
